Switch to KGL

+++ hazel/src/nativeMain/kotlin/hazel/debug/Instrumentor.kt
@@ -9,10 +9,10 @@ import platform.posix.fopen
 import platform.posix.fprintf
 
 data class ProfileResult(
-    val name: String,
-    val start: Long,
-    val end: Long,
-    val threadId: UInt
+	val name: String,
+	val start: Long,
+	val end: Long,
+	val threadId: ULong
 )
 
 data class InstrumentationSession(val name: String)
@@ -23,59 +23,59 @@ private var file: CPointer<FILE>? = null
 
 object Instrumentor {
 
-    fun session(name: String, filepath: String = "profile.json", block: () -> Unit) {
-        beginSession(name, filepath)
-        block()
-        endSession()
-    }
+	fun session(name: String, filepath: String = "profile.json", block: () -> Unit) {
+		beginSession(name, filepath)
+		block()
+		endSession()
+	}
 
-    private fun beginSession(name: String, filepath: String) {
-        if (Hazel.config.isProfileEnabled) {
-            file = fopen(filepath, "w")
-            writeHeader()
-            currentSession = InstrumentationSession(name)
-        }
-    }
+	private fun beginSession(name: String, filepath: String) {
+		if (Hazel.config.isProfileEnabled) {
+			file = fopen(filepath, "w")
+			writeHeader()
+			currentSession = InstrumentationSession(name)
+		}
+	}
 
-    private fun endSession() {
-        if (Hazel.config.isProfileEnabled) {
-            writeFooter()
-            fclose(file)
-            currentSession = null
-            profileCount = 0
-        }
-    }
+	private fun endSession() {
+		if (Hazel.config.isProfileEnabled) {
+			writeFooter()
+			fclose(file)
+			currentSession = null
+			profileCount = 0
+		}
+	}
 
-    fun writeProfile(result: ProfileResult) {
-        file?.let {
-            if (profileCount > 0)
-                fprintf(it, ",")
+	fun writeProfile(result: ProfileResult) {
+		file?.let {
+			if (profileCount > 0)
+				fprintf(it, ",")
 
-            profileCount += 1
+			profileCount += 1
 
-            val name = result.name.replace('"', '\'')
+			val name = result.name.replace('"', '\'')
 
-            fprintf(it, "{")
-            fprintf(it, "\"cat\":\"function\",")
-            fprintf(it, "\"dur\":${result.end - result.start},")
-            fprintf(it, "\"name\":\"$name\",")
-            fprintf(it, "\"ph\":\"X\",")
-            fprintf(it, "\"pid\":0,")
-            fprintf(it, "\"tid\":${result.threadId},")
-            fprintf(it, "\"ts\":${result.start}")
-            fprintf(it, "}")
+			fprintf(it, "{")
+			fprintf(it, "\"cat\":\"function\",")
+			fprintf(it, "\"dur\":${result.end - result.start},")
+			fprintf(it, "\"name\":\"$name\",")
+			fprintf(it, "\"ph\":\"X\",")
+			fprintf(it, "\"pid\":0,")
+			fprintf(it, "\"tid\":${result.threadId},")
+			fprintf(it, "\"ts\":${result.start}")
+			fprintf(it, "}")
 
-            fflush(it)
-        }
-    }
+			fflush(it)
+		}
+	}
 
-    private fun writeHeader() {
-        fprintf(file, "{\"otherData\":{},\"traceEvents\":[")
-        fflush(file)
-    }
+	private fun writeHeader() {
+		fprintf(file, "{\"otherData\":{},\"traceEvents\":[")
+		fflush(file)
+	}
 
-    private fun writeFooter() {
-        fprintf(file, "]}")
-        fflush(file)
-    }
+	private fun writeFooter() {
+		fprintf(file, "]}")
+		fflush(file)
+	}
 }
